Ignore A-R: "arc=none"
When ARC is enabled in Exim, it always adds "arc=none" to the Authentication-Results: header if the message is not ARC signed, along with the spf and dkim results

This causes authres_status to always report an invalid signature.

This patch ignores arc=none, so if the message does not contain any ARC signature but is otherwise valid DKIM etc. It will report valid signature.